pyproject.toml
[project]
name = "haystack-rag"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "fc0597", email = "davide.marchignoli@allitude.it" }
]
dependencies = [
    "haystack-ai>=2.7.0",
    "sentence-transformers>=3.3.1",
    "unstructured-fileconverter-haystack>=0.4.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/haystack_rag"]
